requirements.txt
google-cloud-aiplatform==1.92.0
google-cloud-storage==2.19.0
google-genai==1.14.0
gitpython==3.1.40
google-adk==0.5.0
yfinance==0.2.56
psutil==5.9.5
google-generative-ai==0.8.5
python-dotenv==1.1.0